Move TLight to core/util, add BurstDetector to TBsCostTracker
